Collapse redundant layout-settings route into route-settings
The two-level route indirection only made sense when a pathless layout
wraps routes that don't share a common path prefix (like layout-dashboard
wrapping /dashboard, /sessions/*, /sites/*, etc.). For /settings every
child is under /settings/*, so the parent route itself can own both the
path and the layout component.

- Delete layout-settings/ (settingsLayoutRoute)
- route-settings now renders the SettingsLayout shell with its Outlet
  and redirects bare /settings to /settings/preferences
- route-settings-preferences parent becomes settingsRoute, path nested
